requirements.txt
opencv-python==4.10.0.84
streamlit==1.38.0
numpy
pillow== 10.4.0
tensorflow== 2.18.0
ultralytics== 8.3.87